Fix #7: Clone layer to allow multiple add/remove
Leaflet messes with the internals of layer objects when they're used in a
map, so for the same reason that you have to set up a minimap with a
different layer object than the main map, we have to clone that layer
object internally. This allows you to  use the removeFrom and addTo
functions as much as you like, as demonstrated in the new example.

The clone method is added to L.Util, which might actually be a useful idea
in practice.