Made the map transition too for a v. cool effect
Tested with current versions of IE, Firefox and Chrome. Anything that
doesn't understand transition should just ignore it. (This is also the
case for the transform, worst case the arrow will stop changing for some
where it did before, but damn the torpedoes and steam on ahead! Browsers
will only get more modern.)