Added example: Showing icons in minimap as in #52
This is done by building a LayerGroup, which works out-of-the-box!
Slightly heavy example, works better in Chrome than Firefox, I'm sorry to
admit.

For interactive layers where icons are moved, dded and removed this can probably
be achieved interactively by listening to events on the primary layer and
replicating them on the second layer. (Using the same layer both places
will produce funny effects. Funny as in strange as well as amusing...)